[move]: 도메인 이동 Question, Answer 각각의 도메인에서 qna 도메인으로 이동
